Phase 2: Library refactor
